moving test files and removing /tests/ folder
